My code/video_editor_platform/requirements.txt
# Core Python dependencies for GenAI Video Editor

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Async & Caching
celery==5.3.4
redis==5.0.1
aioredis==2.0.1

# Video Processing
opencv-python==4.8.1.78
moviepy==1.0.3
imageio==2.33.1

# AI/ML Models
torch==2.3.1
torchvision==0.18.1
pillow==10.1.0
numpy==1.26.3

# LLM & APIs (Optional - for paid APIs)
# openai==1.3.8  # Uncomment for OpenAI GPT-4 integration
# elevenlabs==0.2.24  # Uncomment for ElevenLabs TTS
requests==2.31.0
httpx==0.25.2

# Audio Processing
librosa==0.10.0
pydub==0.25.1
scipy==1.11.4
pyttsx3==2.90  # Free text-to-speech (no API required)

# LLM - Open Source (Free)
ollama==0.6.1  # Run local LLaMA-2 models
transformers==4.36.0  # For local inference
accelerate==0.26.0  # Speed up inference

# Cloud Storage
boto3==1.34.0
botocore==1.34.0

# Logging & Monitoring
python-json-logger==2.0.7
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Development
python-dotenv==1.0.0
ipython==8.18.1
